requirements.txt
Flask
Pillow
google-generativeai
